Refactor: Update AppTheme import paths
