+++ app/src/main/java/com/washingtondcsquad/tudee/MainActivity.kt
@@ -1,11 +1,11 @@
 package com.washingtondcsquad.tudee
 
+import SplashScreen
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import com.washingtondcsquad.tudee.presentation.design.AppTheme
-import com.washingtondcsquad.tudee.presentation.design.theme.AppTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -13,8 +13,12 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             AppTheme(
-                useDarkTheme = false,
+                useDarkTheme = true,
             ) {
+                SplashScreen(
+                    title = "Tudee",
+                    isDarkTheme = true
+                )
 
 
             }
