+++ app/src/main/java/com/washingtondcsquad/tudee/presentation/design/textStyle/font.kt
@@ -5,11 +5,6 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import com.washingtondcsquad.tudee.R
 
-val nunito = FontFamily(
-    Font(R.font.nunito, weight = FontWeight.Normal),
-    Font(R.font.nunito_medium, weight = FontWeight.Medium),
-    Font(R.font.nunito_semibold, weight = FontWeight.SemiBold),
-)
 val Cherry_Bomb = FontFamily(
     Font(R.font.cherry_bomb_one, weight = FontWeight.Normal),
 )
